\name{earlywarnings-package}
\alias{earlywarnings-package}
\alias{earlywarnings}
\docType{package}
\title{
Early Warning Signals Toolbox for Detecting Critical Transitions in Timeseries
}
\description{
The Early-Warning-Signals Toolbox provides methods for estimating statistical changes in timeseries that can be used for identifying nearby critical transitions.
}
\details{
\tabular{ll}{
Package: \tab earlywarnings\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-02\cr
}

}
\author{
Vasilis Dakos with contributions from S.C. Carpenter, T. Cline, A. Ives, V. Livina

Maintainer: Vasilis Dakos <vasilis.dakos@gmail.com>
}
\references{
Dakos et al 2012
}
\keyword{ package }
\seealso{
%spatial package -- Optional links to other man pages, e.g. ~~
%\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
%\examples{
%~~ simple examples of the most important functions ~~
%}