% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multimodality_score}
\alias{multimodality_score}
\title{multimodality_score}
\usage{
multimodality_score(dat, detection.threshold = 1, bw.adjust = 1,
  bs.iterations = 100, detection.limit = 1)
}
\arguments{
\item{dat}{Input data matrix (variables x samples)}

\item{detection.threshold}{Mode detection threshold}

\item{bw.adjust}{Bandwidth adjustment}

\item{bs.iterations}{Bootstrap iterations}

\item{detection.limit}{minimum accepted density for a maximum; as a multiple of kernel height}
}
\value{
A list with following elements:
	     score: Fraction of bootstrap samples where multiple modes are observed
	     nmodes: The most frequently observed number of modes in bootrstrap sampling
}
\description{
Calculate multimodality score based on bootstrapped potential analysis
}
\details{
This function repeats potential analysis (Livina et al. 2010) multiple
	    times with bootstrap sampling for each row of the input data
	    (as in Lahti et al. 2014) and returns the specified results
}
\examples{
# Create simulated example data
  #library(microbiome)
  #data(peerj32)
  #multimodality_score(t(peerj32$microbes))
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
Livina et al. (2010). Potential analysis
reveals changing number of climate states during the last 60
kyr. \emph{Climate of the Past}, 6, 77-82.

Lahti et al. (2014). Tipping elements of the human intestinal
ecosystem. \emph{Nature Communications} 5:4344.
}
\keyword{early-warning}

