\name{simulate.gauss}
\alias{simulate.gauss}
\title{simulate method}
\usage{
  \method{simulate}{gauss} (object, nsim = 1, seed = NULL,
    ...)
}
\arguments{
  \item{object}{any gauss-class object}

  \item{...}{additional arguments (currently ignored)}

  \item{nsim}{the number of sims to perform (defaults to
  1)}

  \item{seed}{the random number seed to use}
}
\value{
  a data frame with the time simulated values
}
\description{
  simulate method
}

