\name{simulate}
\alias{simulate}
\title{Simulate generic}
\usage{
  simulate(object, nsim, seed, ...)
}
\arguments{
  \item{object}{any gauss-class object}

  \item{...}{additional arguments (currently ignored)}

  \item{nsim}{the number of sims to perform (defaults to
  1)}

  \item{seed}{the random number seed to use}
}
\value{
  a data frame with the time simulated values
}
\description{
  Simulate generic
}
\details{
  S3 method for simulations
}

