% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YD2PB_grayscale}
\alias{YD2PB_grayscale}
\title{YD2PB_grayscale data set}
\format{TBA}
\source{
TBA
}
\description{
YD2PB_grayscale data set
}
\examples{
#
}
\references{
See citation('earlywarnings')
}
\keyword{datasets}

