% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foldbif}
\alias{foldbif}
\title{foldbif data set}
\format{
TBA
}
\source{
TBA
}
\description{
foldbif data set
}
\examples{
#
}
\references{
See citation('earlywarnings')
}
\keyword{datasetspl}
