\name{PlotPotential}
\alias{PlotPotential}
\title{Description: Plot Potential}
\usage{
  PlotPotential(res, title = "", xlab.text, ylab.text,
    cutoff = 0.5)
}
\arguments{
  \item{res}{output from movpotential function}

  \item{title}{title text}

  \item{xlab.text}{xlab text}

  \item{ylab.text}{ylab text}

  \item{cutoff}{parameter determining the upper limit of
  potential for visualizations

  importFrom akima interp importFrom ggplot2 ggplot

  Returns:}
}
\value{
  \item{ggplot2}{potential plotted}
}
\description{
  Visualization of the potential function from the
  movpotential function
}
\details{
  Arguments:
}
\examples{
#
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  Dakos, V., et al (2012)."Methods for Detecting Early
  Warnings of Critical Transitions in Time Series
  Illustrated Using Simulated Ecological Data." \emph{PLoS
  ONE} 7(7): e41010. doi:10.1371/journal.pone.0041010
}
\keyword{early-warning}

