% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UnivariateGrouping}
\alias{UnivariateGrouping}
\title{Get group assigment indices for univariate data points, given cluster break points}
\usage{
UnivariateGrouping(x, breakpoints)
}
\arguments{
\item{x}{Univariate data vector}

\item{breakpoints}{Cluster breakpoints}
}
\value{
A vector of cluster indices
}
\description{
Get group assigment indices for univariate data points, given cluster break points
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{early-warning}
