\name{UnivariateGrouping}
\alias{UnivariateGrouping}
\title{Description: Get group assigment indices for univariate data points, given cluster break points}
\usage{
UnivariateGrouping(x, breakpoints)
}
\arguments{
  \item{x}{Univariate data vector}

  \item{breakpoints}{Cluster breakpoints

  Returns:}
}
\value{
A vector of cluster indices
}
\description{
Arguments:
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{early-warning}

